refactor(data-space): restructure data-space management by splitting functionality into dedicated modules and removing obsolete code
